S-78: Refactoring purchase table test to common-tests.
