+++ closetr-web/src/test/common-tests.ts
@@ -1,4 +1,5 @@
 import { of } from 'rxjs';
+import { By } from '@angular/platform-browser';
 import {
   clickTest,
   clickAndTestCalledWithMult
@@ -81,4 +82,32 @@ export const getAllClothesComponent = (
   fixture.detectChanges();
   expect(closetService.getAllClothes).toHaveBeenCalledWith(mockUserOne);
   expect(component.closetList).toEqual(mockClosetListRenderedTable);
+}
+
+export const purchaseTableShouldRender = (
+  component, fixture, dateFormatService
+) => {
+  let mockPurchaseTable = {
+    bindBold: "clothingCost",
+    bindRegular: "clothingName",
+    filter: "date",
+    filterBy: "clothingPurchaseDate",
+    filterCriteria: {
+      dateRangeFor: "last month",
+      dateFrom: dateFormatService.dateRangeForFrom("last month"),
+      dateTo: dateFormatService.newDate()
+    },
+    items: mockClosetListRenderedTable
+  };
+  let purchaseTable = fixture.debugElement.query(
+    By.css('#purchase-table')
+  ).componentInstance;
+  component.ngOnInit();
+  fixture.detectChanges();
+  expect(purchaseTable.bindBold).toEqual(mockPurchaseTable.bindBold);
+  expect(purchaseTable.bindRegular).toEqual(mockPurchaseTable.bindRegular);
+  expect(purchaseTable.filter).toEqual(mockPurchaseTable.filter);
+  expect(purchaseTable.filterBy).toEqual(mockPurchaseTable.filterBy);
+  expect(purchaseTable.filterCriteria).toEqual(mockPurchaseTable.filterCriteria);
+  expect(purchaseTable.items).toEqual(mockPurchaseTable.items);
 }
